move edit bookmark functionality into a component
